Adjusted display on devices
I added the viewport meta tag to the html files and played with the text
size on mobile. If you use chromes developer tools and change the device
to a phone or tablet you can see that it fits the screen perfectly on
mobile so you don't have to zoom in.